serverless/requirements.txt
beautifulsoup4
google-cloud-firestore